+++ waspc/src/Wasp/Generator/SdkGenerator/Auth/OAuthAuthG.hs
@@ -8,7 +8,8 @@ import StrongPath (File', Path', Rel', reldir, relfile)
 import qualified StrongPath as SP
 import qualified Wasp.AppSpec.App.Auth as AS.Auth
 import Wasp.Generator.AuthProviders
-  ( discordAuthProvider,
+  ( slackAuthProvider,
+    discordAuthProvider,
     gitHubAuthProvider,
     googleAuthProvider,
     keycloakAuthProvider,
@@ -29,12 +30,14 @@ genHelpers :: AS.Auth.Auth -> Generator [FileDraft]
 genHelpers auth =
   return $
     concat
-      [ [discordHelpers | AS.Auth.isDiscordAuthEnabled auth],
+      [ [slackHelpers | AS.Auth.isSlackAuthEnabled auth],
+        [discordHelpers | AS.Auth.isDiscordAuthEnabled auth],
         [gitHubHelpers | AS.Auth.isGitHubAuthEnabled auth],
         [googleHelpers | AS.Auth.isGoogleAuthEnabled auth],
         [keycloakHelpers | AS.Auth.isKeycloakAuthEnabled auth]
       ]
   where
+    slackHelpers = mkHelpersFd slackAuthProvider [relfile|Slack.tsx|]
     discordHelpers = mkHelpersFd discordAuthProvider [relfile|Discord.tsx|]
     gitHubHelpers = mkHelpersFd gitHubAuthProvider [relfile|GitHub.tsx|]
     googleHelpers = mkHelpersFd googleAuthProvider [relfile|Google.tsx|]